requirements.txt
appdirs==1.4.3
astor==0.8.1
cffi==1.13.2
colorama==0.4.3
esper==1.3
funcparserlib==0.3.6
greenlet==0.4.13
hy==0.18.0
PySDL2==0.9.6
readline==6.2.4.1
rply==0.7.7
